requirements.txt
# This file may be used to create an environment using:
# $ conda create --name <env> --file <this file>
# platform: win-64
bzip2=1.0.8=h2bbff1b_5
ca-certificates=2024.3.11=haa95532_0
colorama=0.4.6=pypi_0
contourpy=1.2.1=pypi_0
cycler=0.12.1=pypi_0
et-xmlfile=1.1.0=pypi_0
filelock=3.13.4=pypi_0
fonttools=4.51.0=pypi_0
fsspec=2024.3.1=pypi_0
jinja2=3.1.3=pypi_0
kiwisolver=1.4.5=pypi_0
libffi=3.4.4=hd77b12b_0
markupsafe=2.1.5=pypi_0
matplotlib=3.8.4=pypi_0
mpmath=1.3.0=pypi_0
networkx=3.3=pypi_0
numpy=1.26.4=pypi_0
openpyxl=3.1.2=pypi_0
openssl=1.1.1w=h2bbff1b_0
packaging=24.0=pypi_0
pandas=2.2.2=pypi_0
pdf2image=1.17.0=pypi_0
pillow=10.3.0=pypi_0
pip=24.0=pypi_0
pyparsing=3.1.2=pypi_0
pypdf2=3.0.1=pypi_0
python=3.10.0=h96c0403_3
python-dateutil=2.9.0.post0=pypi_0
pytz=2024.1=pypi_0
pyyaml=6.0.1=pypi_0
setuptools=69.2.0=pypi_0
six=1.16.0=pypi_0
sqlite=3.41.2=h2bbff1b_0
sympy=1.12=pypi_0
tk=8.6.12=h2bbff1b_0
torch=2.2.2=pypi_0
tqdm=4.66.2=pypi_0
typing-extensions=4.11.0=pypi_0
tzdata=2024.1=pypi_0
vc=14.2=h21ff451_1
vs2015_runtime=14.27.29016=h5e58377_2
wheel=0.43.0=pypi_0
xz=5.4.6=h8cc25b3_0
zlib=1.2.13=h8cc25b3_0
